requirements.txt
altgraph==0.17.5
bottle==0.13.4
cffi==2.0.0
clr_loader==0.2.10
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.2
packaging==26.0
pandas==3.0.0
pefile==2024.8.26
pillow==12.1.0
proxy_tools==0.1.0
pycparser==3.0
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pythonnet==3.0.5
pywebview==6.1
pywin32-ctypes==0.2.3
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.3
